From: "D'Arcy J.M. Cain" <darcy@druid.net> Subject: [HACKERS] backend/utils/adt/nabstime.c
There is a problem with some of the calls to strftime.  The second arg is
missing.  In all cases the buffer is CTZName which, according to the
file init/globals.c, is char CTZName[8] so I have added this value.
I know there should be a #define set up for this but I wasn't sure
which header to put it in.