Compilation of libpq for Win32 breaks on 6.4, because of a change that I missed before the release. It's simply a symbol that is undefined. This patch defines this symbol in "win32.h", so it should have no effect on any other platforms. It should go into 6.4.1 if possible, since compilation is completely broken without it.
I am also attaching a patch for the "win32.mak" file - it leaves a file
behind when doing "make clean" after the library is built on Visual C++ 6.0.
This is not at all as urgent, but I don't see it breaking here, so I think
it might as well go in there too?

//Magnus