> Actually, if you submit a patch that says either "SCROLL is the default" > or "NO SCROLL is the default", it will be rejected as incorrect.  The > reason is that the default behavior is different from either of these, > as is explained in the NOTES section.
Ok, so *that's* where the bit about the query plan being simple enough.
Based on that, ISTM that it should be premissable for us to decide that
a cursor requiring a sort isn't "simple enough" to support SCROLL.

In any case, here's a patch that makes the non-standard behavior easier
for people to find.

Jim C. Nasby