Fix recently-identified PITR recovery hazard: the base backup could contain stale relcache init files (pg_internal.init), and there is no mechanism for updating them during WAL replay.  Easiest solution is just to delete the init files at conclusion of startup, and let the first backend started in each database take care of rebuilding the init file.  Simon Riggs and Tom Lane.
Back-patched to 8.1.  Arguably this should be fixed in 8.0 too, but it would
require significantly more code since 8.0 has no handy startup-time scan of
pg_database to piggyback on.  Manual solution of the problem is possible
in 8.0 (just delete the pg_internal.init files before starting WAL replay),
so that may be a sufficient answer.