Fix up plpgsql's "simple expression" evaluation mechanism so that it behaves safely in the presence of subtransactions.  To ensure that any ExprContext shutdown callbacks are called at the right times, we have to have a separate EState for each level of subtransaction.  Per "TupleDesc reference leak" bug report from Stefan Kaltenbrunner.
Although I'm convinced the code is wrong as far back as 8.0, it doesn't seem
that there are any ways for the problem to really manifest before 8.2: AFAICS,
8.0 and 8.1 only use the ExprContextCallback mechanism to handle set-returning
functions, which cannot usefully be executed in a "simple expression" anyway.
Hence, no backpatch before 8.2 --- the risk of unforeseen breakage seems
to outweigh the chance of fixing something.